Accomodate FEBio 32-bit in time point check

+++ febio.py
@@ -229,7 +229,7 @@ def check_must_points(model, atol=None):
         # Requested and actual times should be treated as half open (,] intervals.
         t0 = t_laststep
         t1 = t_laststep + step.duration
-        t_obs = t[np.logical_and(t > t0, t <= (t1 + np.spacing(t1)))]
+        t_obs = t[np.logical_and(t > t0, t <= (t1 + np.spacing(t1, dtype="float32")))]
         dtmax = step.ticker.dtmax
         t_dtmax = np.array([a for a, b in dtmax.points])
         t_req = t_dtmax[np.logical_and(t_dtmax > t0, t_dtmax <= (t1 + np.spacing(t1)))]
